- datatype 정리  * volume관련 round 발생 문제 -> decimal로 교체

+++ src/Application/DTO/BackTesting/BackTestingRequestDto.cs
@@ -5,14 +5,10 @@ namespace Dreamrosia.Koin.Application.DTO
 {
     public class BackTestingRequestDto : TradingTermsDto
     {
-        public float SeedMoney { get; set; } = DefaultValue.TradingTerms.MaximumAsset4Basic;
-
+        public long SeedMoney { get; set; } = DefaultValue.TradingTerms.MaximumAsset4Basic;
         public DateTime? HeadDate { get; set; } = new DateTime(2017, 09, 25);
-
         public DateTime? RearDate { get; set; } = DateTime.Now.Date;
-
         public bool IncludeOrders { get; set; }
-
         public bool IncludePositons { get; set; }
     }
 }
